+++ src/hooks/LevelInfoLayer.hpp
@@ -4,8 +4,8 @@
 #include <Geode/modify/LevelInfoLayer.hpp>
 #include <sabe.persistenceapi/include/PersistenceAPI.hpp>
 
-extern unsigned long long s_levelInfoLayerDelegate1;
-extern unsigned long long s_levelInfoLayerDelegate2;
+extern size_t s_levelInfoLayerDelegate1;
+extern size_t s_levelInfoLayerDelegate2;
 
 class $modify(PSLevelInfoLayer, LevelInfoLayer) {
 public:
